function parameters explanation
df : is the main raw dataframe off which the derivatives will be created
key_list : is the list of variables that decide the granularity of the dataset. the function is generic enough to create predictors at a store level, store-sku level, sku level, yearly level etc
dim_list : this is the list of variables out of which predictors need to be created. yser may want to add up sales, or inventroy, or stores-sold etc
attr2incement : this is the field that needs to be invremented to create history. so if the model is a yearly model, then this is the year field. more lgoci may be needed by the gran is monghtly or weekly or something like that
yrs : number of history to create.. 4 will give N0, N1, N2, N3 for ex
remove_n0s: n0s are the target year otself.. mostly these will need to go out og the predcictor list.. so this tells the function whether to keep the n0 predictors or not. VERY IMPORTANT. mostly will be 1